Comments corrected

+++ src/logics/transaction_service.py
@@ -10,16 +10,16 @@ class transaction_service:
         Validator.validate_type("transaction_type_element", transaction_type_element, transaction_type)
         self.transaction = getattr(self, transaction_type_element.value.lower())
 
-    def receipt(self, turnover: int, quantity: int) -> int:
+    def receipt(self, turnover: int, quantity: float) -> int:
         Validator.validate_type("turnover", turnover, int)
-        Validator.validate_type("quantity", quantity, int)
+        Validator.validate_type("quantity", quantity, float)
 
         result = turnover + quantity
         return result
     
-    def consumption(self, turnover: int, quantity: int) -> int:
+    def consumption(self, turnover: int, quantity: float) -> int:
         Validator.validate_type("turnover", turnover, int)
-        Validator.validate_type("quantity", quantity, int)
+        Validator.validate_type("quantity", quantity, float)
 
         result = turnover - quantity
         return result
