+++ src/models/transaction.py
@@ -12,7 +12,7 @@ from datetime import datetime
 class transaction_model(base_model_id):
     __storage: storage_model
     __nomenclature: nomenclature_model
-    __quantity: int
+    __quantity: float
     __type_transaction: transaction_type
     __range: range_model
     __period: datetime
@@ -21,7 +21,7 @@ class transaction_model(base_model_id):
     def create(
         storage: storage_model,
         nomenclature: nomenclature_model,
-        quantity: int,
+        quantity: float,
         type_transaction: transaction_type,
         range: range_model,
         period: datetime,
@@ -65,12 +65,12 @@ class transaction_model(base_model_id):
     Количество
     """
     @property
-    def quantity(self) -> int:
+    def quantity(self) -> float:
         return self.__quantity
     
     @quantity.setter
-    def quantity(self, quantity: int):
-        Validator.validate_type("quantity", quantity, int)
+    def quantity(self, quantity: float):
+        Validator.validate_type("quantity", quantity, float)
         Validator.validate_more_permissible_value("quantity", quantity, 0)
         self.__quantity = quantity
